sample-applications/postgresql-python-client/postgresql-python-app/requirements.txt
psycopg2-binary
Faker
